Reorganize file structure to separate cli and server files
