+++ package.json
@@ -11,7 +11,7 @@
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "bin": {
-    "compute-js-static-publish": "./build/cli.js"
+    "compute-js-static-publish": "./build/cli/index.js"
   },
   "repository": {
     "type": "git",
