+++ src/cli/util/hash.ts
