+++ src/server/assets/asset-manager.ts
